UI pet update logic brach moi lan 2
