+++ packages/doctor_pet/lib/common_widget/data_title_widget.dart
@@ -1,4 +1,4 @@
-import 'package:doctor_pet/core/data/DataTitleModel.dart';
+import 'package:doctor_pet/core/data/data_title_model.dart';
 import 'package:flutter/material.dart';
 
 class DataTitleWidget extends StatelessWidget {
@@ -10,7 +10,7 @@ class DataTitleWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return Row(
       children: [
-        const SizedBox(width: 70),
+        const SizedBox(width: 90),
         ...titles
             .map(
               (e) => Flexible(
