+++ packages/doctor_pet/lib/utils/app_enum.dart
@@ -4,4 +4,5 @@ enum Role {
   doctor,
   staff,
   customer,
+
 }
